# contracts/analysis.openapi.yaml
# Generated from diagnostics/analysis (FastAPI) and committed as the contract.
# §10.1: contracts/ is the single source of truth — the frontend's TypeScript types are
# generated from this file, and analysis/tests/test_contract.py fails if the served
# schema and this file disagree, so an API change that was not propagated breaks the
# gate rather than drifting quietly.
#
# Regenerate with: make contract
components:
  schemas:
    ComponentOrigin:
      description: "One as-built component of an assembly, and the supplier lot it was drawn from.\n\n\
        Everything but the serial and the position is nullable, and null means **unknown**, not\nabsent:\
        \ a component named by an assembly whose own read event lies before the gateway's\nhistory horizon\
        \ is one this system knows exists and knows nothing else about. Dropping\nsuch a component would\
        \ make \xA73.5 scenario 7's containment list quietly short, which is\nthe one failure mode that\
        \ matters at the moment a containment list is wanted."
      properties:
        component_serial:
          title: Component Serial
          type: string
        lane:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lane
        lot_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Lot Code
        position:
          title: Position
          type: integer
        read_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Read At
        supplier:
          anyOf:
          - type: string
          - type: 'null'
          title: Supplier
      required:
      - component_serial
      - position
      - lane
      - read_at
      - lot_code
      - supplier
      title: ComponentOrigin
      type: object
    Coverage:
      description: "\xA74.4: without gap markers, missing data is indistinguishable from a quiet machine.\n\
        \n`fully_covered` is carried rather than left to a reader comparing `covered_fraction`\nagainst\
        \ 1.0, for the reason `coverage.Coverage` gives: `gaps == []` and \"fully covered\"\nare only\
        \ the same fact when nothing was clipped into `gaps` in the first place.\n\n`window` is repeated\
        \ here and in the response that embeds this, and deliberately: the\ncoverage of a window is unreadable\
        \ without the window it is of, and a UI that passes\nthis object around alone would otherwise\
        \ be holding a fraction of nothing."
      properties:
        covered_fraction:
          title: Covered Fraction
          type: number
        fully_covered:
          title: Fully Covered
          type: boolean
        gaps:
          items:
            $ref: '#/components/schemas/Gap'
          title: Gaps
          type: array
        observed:
          $ref: '#/components/schemas/ObservedSpan'
        window:
          $ref: '#/components/schemas/Window'
      required:
      - window
      - gaps
      - covered_fraction
      - fully_covered
      - observed
      title: Coverage
      type: object
    DefectClassCount:
      description: "One class, and how many parts in the window scored at or above the threshold.\n\n\
        Not a partition of the rejects: \xA73.4's six scores are independent and do not sum to 1,\nso\
        \ a part the model believes carries two defects counts under both. The counts can\ntherefore total\
        \ more than `rejects`, and that is the answer rather than a rounding of it."
      properties:
        count:
          title: Count
          type: integer
        defect_class:
          title: Defect Class
          type: string
      required:
      - defect_class
      - count
      title: DefectClassCount
      type: object
    Disposition:
      description: How the part left the line, and why. Absent while the part is still on it.
      properties:
        at:
          format: date-time
          title: At
          type: string
        disposition:
          title: Disposition
          type: string
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      required:
      - at
      - disposition
      - reason
      title: Disposition
      type: object
    Gap:
      description: "\xA74.4: without gap markers, missing data is indistinguishable from a quiet machine."
      properties:
        from_ts:
          format: date-time
          title: From Ts
          type: string
        reason:
          title: Reason
          type: string
        to_ts:
          format: date-time
          title: To Ts
          type: string
      required:
      - from_ts
      - to_ts
      - reason
      title: Gap
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Inspection:
      description: "\xA73.4's verdict for this part, or absent if the part has not been inspected.\n\n\
        `defect_classes` and `confidences` are parallel arrays over **every** class the\nclassifier scores,\
        \ on good parts too \u2014 a good part is six low scores, not an absent\nvector. `confidence`\
        \ is the confidence in the OK/NOK verdict and is not one of them;\nreading the vector as a distribution\
        \ is the measured defect that once reported a good\npart as 27 % confident and ~30 % misaligned.\n\
        \nBoth arrays are nullable because M1 wrote rows before the vector existed. Null is\n\"this row\
        \ predates the widened event\", not \"this part scored nothing\"."
      properties:
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        confidences:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Confidences
        defect_classes:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Defect Classes
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        model_version:
          title: Model Version
          type: string
        result:
          title: Result
          type: string
        source_ts:
          format: date-time
          title: Source Ts
          type: string
        station:
          title: Station
          type: string
      required:
      - source_ts
      - station
      - result
      - defect_classes
      - confidences
      - confidence
      - model_version
      - image_url
      title: Inspection
      type: object
    InspectionStats:
      description: "Counts over a window, with everything needed to tell a real answer from an empty one.\n\
        \n`by_defect_class` and `rejects_without_class` together account for every reject, and\nneither\
        \ is readable without the other. They do not sum to `rejects` \u2014 a part scoring\nhigh on two\
        \ classes appears twice in the breakdown \u2014 but `rejects_without_class` is\nexactly the part\
        \ of `rejects` the breakdown cannot explain, so an empty breakdown beside\na non-zero `rejects`\
        \ now says *which* it is: 30 rejects and 30 unclassified is a\nmeasurement, and 30 rejects and\
        \ 0 unclassified with an empty breakdown is impossible."
      properties:
        by_defect_class:
          items:
            $ref: '#/components/schemas/DefectClassCount'
          title: By Defect Class
          type: array
        coverage:
          $ref: '#/components/schemas/Coverage'
        defect_class_threshold:
          title: Defect Class Threshold
          type: number
        group_by:
          anyOf:
          - enum:
            - time
            - carrier
            - lane
            - defect_class
            type: string
          - type: 'null'
          title: Group By
        groups:
          anyOf:
          - items:
              $ref: '#/components/schemas/StatsGroup'
            type: array
          - type: 'null'
          title: Groups
        rejects:
          title: Rejects
          type: integer
        rejects_without_class:
          title: Rejects Without Class
          type: integer
        sample_serials:
          items:
            type: string
          title: Sample Serials
          type: array
        total:
          title: Total
          type: integer
        window:
          $ref: '#/components/schemas/Window'
      required:
      - window
      - total
      - rejects
      - by_defect_class
      - defect_class_threshold
      - rejects_without_class
      - sample_serials
      - coverage
      - group_by
      - groups
      title: InspectionStats
      type: object
    ObservedSpan:
      description: "What the window actually holds, across every stream the analysis reads.\n\nThis is\
        \ the half of coverage the gap rows cannot supply. A window with no gaps and no\nrows is a quiet\
        \ line; a window with no rows because a gap covers all of it is a\nblackout; and the two are the\
        \ same empty answer until something counts what is there.\n`events` is deliberately a total over\
        \ the streams rather than a breakdown \u2014 the\nquestion it answers is \"is there anything here\
        \ at all\", and a per-stream count would\ninvite it being read as a production figure, which it\
        \ is not."
      properties:
        events:
          title: Events
          type: integer
        from_ts:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: From Ts
        to_ts:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: To Ts
      required:
      - from_ts
      - to_ts
      - events
      title: ObservedSpan
      type: object
    Part:
      description: "\xA714's end-to-end trace of one serial, every section read by that serial alone.\n\
        \nA section that is empty or null is a section this system has no row for, and the part\nis answered\
        \ anyway. The two shapes that produces are both ordinary rather than\nexceptional: a part between\
        \ S2 and S3 has no verdict yet, and an assembly created\nbefore the gateway's history horizon\
        \ has no creation instant, no carrier and no\ngenealogy, because the one event carrying all three\
        \ arrived before the gateway did."
      properties:
        assembly_serial:
          title: Assembly Serial
          type: string
        carrier_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Carrier Id
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Created At
        disposition:
          anyOf:
          - $ref: '#/components/schemas/Disposition'
          - type: 'null'
        genealogy:
          items:
            $ref: '#/components/schemas/ComponentOrigin'
          title: Genealogy
          type: array
        inspection:
          anyOf:
          - $ref: '#/components/schemas/Inspection'
          - type: 'null'
        process_curves:
          items:
            $ref: '#/components/schemas/ProcessCurve'
          title: Process Curves
          type: array
        process_values:
          items:
            $ref: '#/components/schemas/ProcessValue'
          title: Process Values
          type: array
      required:
      - assembly_serial
      - created_at
      - carrier_id
      - genealogy
      - process_values
      - process_curves
      - inspection
      - disposition
      title: Part
      type: object
    ProcessCurve:
      description: "D6's force\u2013distance curve for this part, as the samples the press recorded.\n\
        \nThe thing the two scalars cannot reconstruct: two presses reach the same peak at the\nsame final\
        \ position by different routes, and the route is the diagnosis (\xA73.4a)."
      properties:
        samples:
          items:
            type: number
          title: Samples
          type: array
        signal:
          title: Signal
          type: string
        station:
          title: Station
          type: string
      required:
      - station
      - signal
      - samples
      title: ProcessCurve
      type: object
    ProcessValue:
      description: 'One number a station recorded against this serial at the instant of production.


        The signal is the name the plant''s own event carries (`PeakForce`), which is not

        always the name of the historised stream beside it (`JoiningForcePeak`). Nothing

        between the plant and here translates the two, so nothing can translate them wrongly.'
      properties:
        signal:
          title: Signal
          type: string
        station:
          title: Station
          type: string
        value:
          title: Value
          type: number
      required:
      - station
      - signal
      - value
      title: ProcessValue
      type: object
    StatsGroup:
      description: "One value of the requested grouping, and the reject share within it.\n\n**`parts`\
        \ is the denominator and it is not the same quantity for every grouping.** For\n`time`, `carrier`\
        \ and `lane` the group is a set of parts and `parts` counts them. For\n`defect_class` the group\
        \ is a *class*, which is not a set of parts at all \u2014 \xA73.4's six\nscores are independent,\
        \ a part can carry several and most carry none \u2014 so the\ndenominator is every inspected part\
        \ in the window, the same number under every class,\nand `rejects` is how many of them scored\
        \ at or above the threshold on this one.\n\nThat choice is the per-part denominator, and it is\
        \ the same one `/inspection/patterns`\ntests with: a part is one trial for each class, not one\
        \ trial shared between them. The\nconsequence to read the numbers with is that the defect-class\
        \ shares do not sum to the\nwindow's reject rate and were never going to \u2014 a part carrying\
        \ two classes is counted\nunder both, exactly as `DefectClassCount` describes.\n\n`lane` is a\
        \ third case again, and the one to be careful with: every assembly draws one\ncomponent from *each*\
        \ feeder lane (\xA73.5), so the lane groups hold the same parts and\ntheir `parts` counts sum\
        \ to more than the window's. The counts are real; what they do\nnot support is a comparison between\
        \ the lanes, which is why `/inspection/patterns`\nrefuses that dimension rather than returning\
        \ a verdict on it.\n\n`from_ts` and `to_ts` are set for `time` and null for everything else. They\
        \ are the\nbucket's edges **clipped to the requested window**, so a bucket the window only partly\n\
        covers is visibly short rather than quietly under-counted."
      properties:
        from_ts:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: From Ts
        key:
          title: Key
          type: string
        parts:
          title: Parts
          type: integer
        reject_share:
          anyOf:
          - type: number
          - type: 'null'
          title: Reject Share
        rejects:
          title: Rejects
          type: integer
        to_ts:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: To Ts
      required:
      - key
      - from_ts
      - to_ts
      - parts
      - rejects
      - reject_share
      title: StatsGroup
      type: object
    TimeResolution:
      description: 'What `/time/resolve` made of a phrase, and as of when.


        `now` is echoed because `closed` is a statement about the clock and not about the

        window: the same expression resolved a minute later can answer the same window and a

        different `closed`, and a cache that kept the flag without the instant it was taken at

        would be asserting a fact it cannot support.'
      properties:
        closed:
          title: Closed
          type: boolean
        expression:
          title: Expression
          type: string
        label:
          title: Label
          type: string
        now:
          format: date-time
          title: Now
          type: string
        window:
          $ref: '#/components/schemas/Window'
      required:
      - expression
      - window
      - label
      - closed
      - now
      title: TimeResolution
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    Window:
      properties:
        from_ts:
          format: date-time
          title: From Ts
          type: string
        to_ts:
          format: date-time
          title: To Ts
          type: string
      required:
      - from_ts
      - to_ts
      title: Window
      type: object
info:
  title: machine-agent analysis
  version: 0.1.0
openapi: 3.1.0
paths:
  /coverage:
    get:
      description: Where the data is, and where it is not, over a half-open window.
      operationId: coverage
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Coverage'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Coverage
  /inspection/stats:
    get:
      description: "Counts over a half-open window, with the gaps that make them incomplete.\n\n**The\
        \ breakdown reads \xA73.4's score vector, not a scalar class.** It grouped by\n`inspection_results.defect_class`\
        \ until M2b Task 7, and by then nothing filled that\ncolumn: the plant sends `DefectClasses`,\
        \ the gateway's select clause named the old\nscalar, asyncua maps a select clause it cannot resolve\
        \ to `Variant(None)` rather than\nerroring, the writer resolves that to `DBNull`, and `defect_class\
        \ IS NOT NULL` then\nemptied the group-by. No exception and no 500 \u2014 `by_defect_class: []`\
        \ beside a correct\n`total` and `rejects`, which is a silently empty answer to \"which defects\
        \ are we\nseeing\" and the exact failure \xA71 exists to prevent. The column is no longer in the\
        \ read\nlayer at all, so the query that produced that answer can no longer be written.\n\n`group_by`\
        \ is absent by default and `groups` is then null rather than empty: a caller\nthat asked for no\
        \ grouping and a caller that asked for one over an empty window are not\nentitled to the same\
        \ answer."
      operationId: inspectionStats
      parameters:
      - in: query
        name: group_by
        required: false
        schema:
          anyOf:
          - enum:
            - time
            - carrier
            - lane
            - defect_class
            type: string
          - type: 'null'
          title: Group By
      - in: query
        name: from
        required: true
        schema:
          format: date-time
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date-time
          title: To
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionStats'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Inspection Stats
  /parts/{serial}:
    get:
      description: "\xA714's trace: genealogy, what each station recorded, the verdict and the disposition.\n\
        \n**Every statement below is keyed by the serial and nothing else.** \xA73.4a: the per-part\n\
        record is authoritative for the part and is never reconstructed by joining the time\nseries on\
        \ \"which part was at S2 at 02:14:07\" \u2014 with buffers, variable takt and history\ngaps that\
        \ association is an inference, and inferred traceability is what makes a\ncontainment list unusable\
        \ at the moment it matters. There is no time range anywhere in\nthis function, and `tests/test_traceability.py`\
        \ is what fails if one appears.\n\n\"Station history\" here is what happened to the part at each\
        \ station \u2014 S1's creation\ninstant, S2's press record, S3's verdict, S4's disposition \u2014\
        \ and not\n`part_station_events`, which \xA75.2 says has no source: no event the plant publishes\n\
        carries a station entry or exit instant, and writing the *processing* instant into\n`entered_at`\
        \ would be exactly the quiet wrong answer this system refuses."
      operationId: getPart
      parameters:
      - in: path
        name: serial
        required: true
        schema:
          title: Serial
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Part'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Part
  /parts/{serial}/image:
    get:
      description: "Rejects only. A good part has no image and that is not a missing value (\xA73.4)."
      operationId: getPartImage
      parameters:
      - in: path
        name: serial
        required: true
        schema:
          title: Serial
          type: string
      responses:
        '200':
          content:
            image/png: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Part Image
  /time/resolve:
    get:
      description: 'Resolve a fixed time phrase to a concrete UTC window against the shift calendar.


        **An expression this does not understand is a 422 carrying the list of ones it does,

        never a guess.** A near-miss guessed wrong is the worst of the three possible outcomes:

        the caller gets a window, believes it is the one it asked for, and every number computed

        over it is an answer to a different question. The list is in the error so that the agent

        can retry with something real instead of rephrasing at random.'
      operationId: resolveTime
      parameters:
      - in: query
        name: expression
        required: true
        schema:
          title: Expression
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeResolution'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resolve Time